% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_client.R
\docType{class}
\name{az_client}
\alias{az_client}
\title{AzureRMR client class}
\format{An R6 object of class \code{az_rm}.}
\usage{
az_client
}
\description{
Container class for interacting with Azure Resource Manager and Azure AD Graph.
}
\section{Methods}{

This class exposes the methods of the underlying \code{az_rm} and \code{az_graph} classes, via active bindings. See the documentation for those classes. It includes one additional method:

\code{refresh()}: Refreshes the tokens for the \code{az_rm} and \code{az_graph} component objects.
}

\section{Initialization}{

The recommended way to create new instances of this class is via the \link{create_azure_login} and \link{get_azure_login} functions.
}

\seealso{
\link{create_azure_login}, \link{get_azure_login}, \link{az_graph}, \link{az_rm}

\href{https://docs.microsoft.com/en-us/azure/azure-resource-manager/resource-group-overview}{Azure Resource Manager overview},
\href{https://docs.microsoft.com/en-us/rest/api/resources/}{REST API reference}

\href{https://docs.microsoft.com/en-us/azure/active-directory/develop/active-directory-graph-api}{Azure AD Graph overview},
\href{https://docs.microsoft.com/en-au/previous-versions/azure/ad/graph/api/api-catalog}{REST API reference}
}
\keyword{datasets}
