% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_login.R
\name{create_az_login}
\alias{create_az_login}
\alias{get_az_login}
\alias{delete_az_login}
\alias{list_az_logins}
\alias{refresh_az_logins}
\title{Functions to login to Azure Resource Manager}
\usage{
create_az_login(tenant, app = NULL, auth_type = "device_code", ...)

get_az_login(tenant, refresh = TRUE)

delete_az_login(tenant, confirm = TRUE)

list_az_logins()

refresh_az_logins()
}
\arguments{
\item{tenant}{The Azure Active Directory tenant for which to obtain a login client. Can be a name ("mytenant"), a fully qualified domain name ("mytenant.microsoft.com"), or a GUID.}

\item{app}{The app ID to authenticate with. Defaults to the app used by the Azure CLI.}

\item{auth_type}{The type of authentication to use. Can be either "device_code" or "client_credentials". Use the latter if you supply a password.}

\item{...}{Other arguments to pass to \code{az_rm$new()}.}

\item{refresh}{For \code{get_az_login}, whether to refresh the authentication token on loading the client.}

\item{confirm}{For \code{delete_az_login}, whether to ask for confirmation before deleting.}
}
\value{
For \code{create_az_login} and \code{get_az_login}, an object of class \code{az_rm}, representing the ARM client. For \code{list_az_logins}, a list of such objects.
}
\description{
Functions to login to Azure Resource Manager
}
\details{
These functions allow you to login to Azure Resource Manager (ARM).
\itemize{
\item \code{create_az_login} will create a login client. R will display a code and prompt you to visit the Microsoft login page in your browser. You then enter the code along with your Azure Active Directory credentials, which completes the authentication process. You only have to create a login client once per tenant: the resulting ARM client object is saved on your machine and reused automatically in subsequent R sessions.
\item \code{get_az_login} will load a previously saved ARM client object for the given tenant. If this is the first time you are logging in for this tenant, the client object is created.
\item \code{delete_az_login} deletes the client object for the given tenant from your machine. Note that this doesn't invalidate any existing client you may have in your R session.
\item \code{list_az_logins} lists client objects that have been previously saved.
\item \code{refresh_az_logins} refreshes all client objects existing on your machine.
}

\code{create_az_login} is roughly equivalent to the Azure CLI command \code{az login} with no arguments, and in fact uses the same app ID as the CLI by default.
}
\examples{
\dontrun{

# this will create a Resource Manager client for the AAD tenant 'microsoft.onmicrosoft.com'
# only has to be run once per tenant
az <- create_az_login("microsoft")

# in subsequent sessions, you can retrieve the client without re-authenticating
# authentication details will automatically be refreshed
az <- get_az_login("microsoft")

# refresh (renew) authentication details for clients for all tenants
refresh_az_logins()

}
}
\seealso{
\link{az_rm}, \href{https://docs.microsoft.com/en-us/cli/azure/?view=azure-cli-latest}{Azure CLI documentation}
}
