% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureRMR.R
\name{AzureRMR_dir}
\alias{AzureRMR_dir}
\title{Data directory for AzureRMR}
\usage{
AzureRMR_dir()
}
\value{
A string containing the data directory.
}
\description{
Data directory for AzureRMR
}
\details{
AzureRMR stores authentication credentials and OAuth tokens in a user-specific directory, using the rappdirs package. On recent Windows versions, this will usually be in the location \code{C:\\Users\\(username)\\AppData\\Local\\AzureR\\AzureRMR}. On Linux, it will be in \code{~/.local/share/AzureRMR}, and on MacOS, it will be in \code{~/Library/Application Support/AzureRMR}. The working directory is not touched (which significantly lessens the risk of accidentally introducing cached tokens into source control).
}
\seealso{
\link{get_azure_token}, \link{get_azure_login}
}
