% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_azure_rm.R
\name{call_azure_rm}
\alias{call_azure_rm}
\alias{call_azure_url}
\title{Call the Azure Resource Manager REST API}
\usage{
call_azure_rm(token, subscription, operation, ..., options = list(),
  api_version = getOption("azure_api_version"))

call_azure_url(token, url, ..., http_verb = c("GET", "DELETE", "PUT", "POST",
  "HEAD", "PATCH"), http_status_handler = c("stop", "warn", "message",
  "pass"), auto_refresh = TRUE)
}
\arguments{
\item{token}{An Azure OAuth token, of class \link{AzureToken}.}

\item{subscription}{A subscription ID.}

\item{...}{Other arguments passed to lower-level code, ultimately to the appropriate functions in httr.}

\item{operation:}{The operation to perform, which will form part of the URL path.}

\item{options:}{A named list giving the URL query parameters.}

\item{api_version:}{The API version to use, which will form part of the URL sent to the host.}

\item{url:}{A complete URL to send to the host.}

\item{http_verb:}{The HTTP verb as a string, one of \code{GET}, \code{PUT}, \code{POST}, \code{DELETE}, \code{HEAD} or \code{PATCH}.}

\item{http_status_handler:}{How to handle in R the HTTP status code of a response. \code{"stop"}, \code{"warn"} or \code{"message"} will call the appropriate handlers in httr, while \code{"pass"} ignores the status code.}

\item{auto_refresh:}{Whether to refresh/renew the OAuth token if it is no longer valid.}
}
\value{
If \code{http_status_handler} is one of \code{"stop"}, \code{"warn"} or \code{"message"}, the status code of the response is checked and if an error is not thrown, the content of the response is returned with the status code attached as the "status" attribute. If \code{http_status_handler} is \code{"pass"}, the entire response is returned without modification.
}
\description{
Call the Azure Resource Manager REST API
}
\details{
These functions form the low-level interface between R and Resource Manager. \code{call_azure_rm} builds a URL from its arguments and passes it to \code{call_azure_url}. Authentication is handled automatically.
}
\seealso{
\link{httr:GET}, \link{httr:stop_for_status}, \link{httr:content}
}
