% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_subscription}
\alias{is_subscription}
\alias{is_resource_group}
\alias{is_resource}
\alias{is_template}
\alias{is_azure_token}
\title{Informational functions}
\usage{
is_subscription(object)

is_resource_group(object)

is_resource(object)

is_template(object)

is_azure_token(object)
}
\arguments{
\item{object}{An R object.}
}
\value{
A boolean.
}
\description{
These functions return whether the object is of the corresponding AzureRMR class.
}
