% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureToken.R
\name{normalize_tenant}
\alias{normalize_tenant}
\title{Normalizes a tenant}
\usage{
normalize_tenant(tenant)
}
\arguments{
\item{tenant}{An Azure Active Directory tenant. This can be a name ("myaadtenant"), a fully qualified domain name ("myaadtenant.onmicrosoft.com" or "mycompanyname.com"), or a valid GUID.}
}
\value{
The normalized ID or name of the tenant.
}
\description{
Normalizes a tenant
}
\details{
This function is used by \code{get_azure_token} to recognise a tenant input. A tenant can be identified either by a GUID, or its name, or a fully-qualified domain name (FQDN). The rules for normalizing a tenant are:
\enumerate{
\item If \code{tenant} is a valid GUID, return its canonically formatted value
\item Otherwise, if it is a FQDN, return it
\item Otherwise, if it is not the string "common", append ".onmicrosoft.com" to it
\item Otherwise, return the value of \code{tenant}
}
}
\seealso{
\link{get_azure_token},

\href{https://docs.microsoft.com/en-us/dotnet/api/system.guid.parse]}{Parsing rules for GUIDs in .NET}. \code{normalize_tenant} recognises the "N", "D", "B" and "P" formats for GUIDs.
}
