% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\name{init_pool}
\alias{init_pool}
\alias{delete_pool}
\title{Manage parallel Azure connections}
\usage{
init_pool(size = 10, restart = FALSE, ...)

delete_pool()
}
\arguments{
\item{size}{The number of background R processes to create. Limit this is you are low on memory.}

\item{restart}{For \code{init_pool}, whether to terminate an already running pool first.}

\item{...}{Other arguments passed on to \code{parallel::makeCluster}.}
}
\description{
Manage parallel Azure connections
}
\details{
AzureRMR provides the ability to parallelise communicating with Azure by utilizing a pool of R processes in the background. This often leads to major speedups in scenarios like downloading large numbers of small files, or communicating with a cluster of virtual machines. The pool is created by calling \code{init_pool}. It remains persistent for the session or until terminated by \code{delete_pool}.

If \code{init_pool} is called and the current pool is smaller than \code{size}, it is resized.
}
