% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport_AzureAuth.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{clean_token_directory}
\alias{delete_azure_token}
\alias{get_azure_token}
\alias{is_azure_token}
\alias{is_azure_v1_token}
\alias{is_guid}
\alias{list_azure_tokens}
\alias{AzureR_dir}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AzureAuth}{\code{\link[AzureAuth]{AzureR_dir}}, \code{\link[AzureAuth:get_azure_token]{clean_token_directory}}, \code{\link[AzureAuth:get_azure_token]{delete_azure_token}}, \code{\link[AzureAuth]{get_azure_token}}, \code{\link[AzureAuth:get_azure_token]{is_azure_token}}, \code{\link[AzureAuth:get_azure_token]{is_azure_v1_token}}, \code{\link[AzureAuth:get_azure_token]{is_azure_v1_token}}, \code{\link[AzureAuth:guid]{is_guid}}, \code{\link[AzureAuth:get_azure_token]{list_azure_tokens}}}
}}

